Fixes lowercase filename issue for linux
