+++ public/index.html
@@ -30,5 +30,5 @@
       //     }
       // })();
     </script>
-  <script type="text/javascript" src="bundle.56ebec82cfec7dc7dea1.js"></script></body>
+  <script type="text/javascript" src="bundle.a432228cd50ad5f80ae3.js"></script></body>
 </html>
