+++ src/actions/index.js
@@ -1,4 +1,4 @@
 export * from './settings.js';
 //  TODO: lets make videos just import them all?
-export * from './videosFetch.js';
-export * from './videosPlayer.js';
+export * from './videos-fetch.js';
+export * from './videos-player.js';
