--- src/components/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-// import Header from './Header/Header.js';
-import Player from './Player/Player.js';
-
-const App = () => (
-  // <Header />
-  <div className="app-l">
-    <Player />
-  </div>
-);
-
-export default App;
