+++ src/components/app.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Player from './player/player.js';
+
+const App = () => (
+    <div className="app-l">
+        <Player />
+    </div>
+);
+
+export default App;
