+++ src/reducers/videos-player.js
